% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{drawTree}
\alias{drawTree}
\title{Draw a tree from a newick string; add to current plot}
\usage{
drawTree(tree, col = "black", lwd = 1, timeScale = 1,
  call.plotSpr = FALSE, cex.leafname = 2, leafCol = col,
  leafLabels = NULL, mod = NULL)
}
\arguments{
\item{tree}{A newick string representing a tree}

\item{col}{Either a character string (single value) giving color of tree. Or, a
list giving color to plot each leaf branch. Internal branches will
be plotted as "average" color among child leafs.}

\item{lwd}{Line width}

\item{timeScale}{multiply all branches by this value}

\item{call.plotSpr}{If TRUE and if a SPR event is encoded in the tree (with NHX tags),
then draw the SPR event on the tree (currently only works if mod is not NULL)}

\item{cex.leafname}{Character size of leaf names}

\item{leafCol}{If given, this can be a list assigning a color to each leaf name
(individual names also work if haploid names are in the form indName_1 and indName_2)}

\item{leafLabels}{A list indexed by the leaf names in the tree, giving the label to display
for each leaf name. If NULL, the leaf names are displayed.}

\item{mod}{(Advanced; for use with multi-population version of ARGweaver) A model
file read in with the function readPopModel, will draw population model underneath tree}
}
\description{
Draw a tree from a newick string; add to current plot
}
\note{
The parameters col, leafCol, and leafLabels should all be lists indexed by leaf
names of the tree. However, they can also be diploid individual names, if the leafs
are named with the convention indName_1 and indName_2
}
\seealso{
\code{\link{plotTrees}} for function which creates a new plot
}
