% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArgSummary.R
\name{getCols}
\alias{getCols}
\title{Choose columns from summary file}
\usage{
getCols(x, inds, prefix = "", suffix = "", or = FALSE)
}
\arguments{
\item{x}{A data frame read by readSummary()}

\item{inds}{A vector of character strings containing individual
(or haplotype names) relevant to desired columns}

\item{prefix}{Choose only columns with this prefix}

\item{suffix}{Choose only columns with this suffix}

\item{or}{If TRUE, return columns that mention any of the individuals.
If FALSE, return columns that mention all the individuals}
}
\value{
A list of column names from x that meet the criteria
}
\description{
Given an ARGweaver summary file,
return columns of interest
}
\note{
This may not work as desired if some individual names are subsets of others
}
