% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{getHaps}
\alias{getHaps}
\title{Get haplotype names from individual names}
\usage{
getHaps(inds)
}
\arguments{
\item{inds}{character string of diploid individual names}
}
\value{
character vector twice as long as inds, with "_1" and "_2" appended
to each individual name
}
\description{
Get haplotype names from individual names
}
