% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimes.R
\name{getLogFile}
\alias{getLogFile}
\title{Determine argweaver log file starting with file name}
\usage{
getLogFile(filename)
}
\arguments{
\item{filename}{Beginning of log filename}
}
\value{
name of ARGweaver log file if there is one good candidate, otherwise error
}
\description{
Determine argweaver log file starting with file name
}
