% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{getOuterPopPath}
\alias{getOuterPopPath}
\title{Get the population path of the parent branch of tree}
\usage{
getOuterPopPath(tree)
}
\arguments{
\item{tree}{A character vector with a newick tree}
}
\value{
The population path encoded in the newick string for outermost branch
}
\description{
Get the population path of the parent branch of tree
}
\note{
The population path is usually zero; it is only non-zero for the multiple-population
ARGweaver model
}
