% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popModel.R
\name{getPathsFromLogFile}
\alias{getPathsFromLogFile}
\title{Get population paths from ARGweaver log file}
\usage{
getPathsFromLogFile(logfile, ntimes)
}
\arguments{
\item{logfile}{The log file name}

\item{ntimes}{The number of discrete times}
}
\value{
A data frame with one row for each path; each column is a time interval
}
\description{
Get population paths from ARGweaver log file
}
