% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{getRootNode}
\alias{getRootNode}
\title{Get the index of the root node for a tree}
\usage{
getRootNode(tree)
}
\arguments{
\item{tree}{An object of type "phylo" (a tree loaded in ape).}
}
\value{
The index of the root node in the tree object
}
\description{
Get the index of the root node for a tree
}
