% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimes.R
\name{getTimes}
\alias{getTimes}
\title{Compute log-distributed discrete times used by arg-sample}
\usage{
getTimes(n = 20, maxTime = 2e+05, delta = 0.01)
}
\arguments{
\item{n}{The number of time points between 0 and maxTime, inclusive}

\item{maxTime}{The maximum time}

\item{delta}{Spacing parameter. Times get denser towards t=0
as delta gets small}
}
\value{
A vector of times of length n
}
\description{
This function returns a vector of times which are spaced
on a log scale according to the parameter delta.
}
