% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTimes.R
\name{getTimesFromLogFile}
\alias{getTimesFromLogFile}
\title{Get times from an argweaver log file}
\usage{
getTimesFromLogFile(logfile)
}
\arguments{
\item{logfile}{the .log file output by ARGweaver}
}
\value{
Numeric list of discrete times
}
\description{
Get times from an argweaver log file
}
