% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{getTreeDepth}
\alias{getTreeDepth}
\title{Get maximum distance from a node to the leaf}
\usage{
getTreeDepth(tree, idx = -1)
}
\arguments{
\item{tree}{A tree, either of type "phylo" (read in by ape), or a character
string with a Newick tree}

\item{idx}{The index of the root node. Usually this will be -1 and the true root will
be used. The value of idx refers to the row of the tree$edge data frame,
for trees read by "ape"}
}
\value{
The maximum distance from the indexed node (or root if idx=-1) to any leaf node.
}
\description{
Get maximum distance from a node to the leaf
}
