% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{makeTransparent}
\alias{makeTransparent}
\title{Make a color transparent}
\usage{
makeTransparent(col, alpha = 0.5)
}
\arguments{
\item{col}{The color (as a character vector)}

\item{alpha}{The degree of transparency (0=invisible, 1=solid).}
}
\value{
A character vector with a transparent color matching original color in RGB values
}
\description{
Make a color transparent
}
