% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{meanColor}
\alias{meanColor}
\title{Average color}
\usage{
meanColor(cols)
}
\arguments{
\item{cols}{A list of colors}
}
\value{
A color created by averaging all the R, G, B values of the input colors
}
\description{
Average color
}
