% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiHist.R
\name{multiHist}
\alias{multiHist}
\title{Overlay several histograms}
\usage{
multiHist(x, stat, col = NULL, xlab = stat, breaks = 50, main = "",
  xlim = NULL, ...)
}
\arguments{
\item{x}{list of data frames. Each should have a column named stat}

\item{stat}{gives the column to plot. Should be single character string.}

\item{col}{A list or character vectors of colors to use for each histogram.
If NULL, will use a rainbow color, but each color will be given transparency
value alpha=0.5}

\item{xlab}{Label for x axis}

\item{breaks}{Passed on to hist function. Can be a number of breaks, or a numeric
vector of break intervals.}

\item{main}{Title for plot}

\item{xlim}{limits for x-axis}

\item{...}{Passed to "hist" function}
}
\description{
x should be a list of data frames with the same structure.
Will plot a histogram of the column named by "stat"
for each item in x
}
