% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafTrace.R
\name{plotLeafTrace}
\alias{plotLeafTrace}
\title{Create a leafTrace plot}
\usage{
plotLeafTrace(file = NULL, x = NULL, col = "black", xlim = NULL,
  ylim = NULL, xlab = "Coordinate", ylab = "", add = FALSE,
  subset = NULL, lwd = 1, ...)
}
\arguments{
\item{file}{A leaf trace file (produced by the script arg-layout)}

\item{x}{(alternative to file) A leaf trace data frame read in by readLeafTrace}

\item{col}{Either a character string giving the color of the plot, or a
list indexed by leaf or diploid individual names giving the color for
each leaf/individual. Will assume that leafs are named with the
convention <ind>_1 and <ind>_2 for the two diploid lineages}

\item{xlim}{The range for x coordinates}

\item{ylim}{The range for y coordinates}

\item{xlab}{Label for the x-axis}

\item{ylab}{Label for the y-axis}

\item{add}{If TRUE, add plot to current plot}

\item{subset}{If NULL, plot all leafs. Otherwise only plot individuals/leafs
in this character vector}

\item{lwd}{Line width}

\item{...}{Additional objects passed to the plot function}
}
\description{
The leaftrace draws a horizontal line for every leaf across the ARG,
such that the distance between adjacent lines reflects the distance between
the leaves corresponding to those lines. There is an randomness to the
leaf order and the distance between non-adjacent lines cannot be ascertained
}
