% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmc.R
\name{plotPsmc}
\alias{plotPsmc}
\title{Plot PSMC output file}
\usage{
plotPsmc(psmcFile, mu = 1.4e-08, s = 100, sampleAge = 0,
  add = FALSE, timeScale = 1, xlim = NULL, ylim = NULL,
  xlab = "Time", ylab = "Popsize", log = "x", ...)
}
\arguments{
\item{psmcFile}{file name containing PSMC output}

\item{mu}{Mutation rate per base pair per generation}

\item{s}{The site compression rate used in PSMC}

\item{sampleAge}{(For ancient samples) the sample age, in generations}

\item{add}{If TRUE, add to current plotting device}

\item{timeScale}{Scale all times by this factor (i.e. to convert generations to years)}

\item{xlim}{x-axis range. If NULL, use range in data}

\item{ylim}{y-axis range. If NULL, use range in data}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}

\item{log}{Which axis to use log scale for (either "x", "y", "xy", "n")}

\item{...}{Other arguments to be passed to "segments" function (col, lty, lwd, etc)}
}
\description{
Plot PSMC output file
}
