% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSites.R
\name{plotSites}
\alias{plotSites}
\title{Plot a sites file}
\usage{
plotSites(x = NULL, file = NULL, start = NULL, end = NULL,
  majColor = "black", minColor = "red", stretch = 1, indOrder = NULL,
  textColor = NULL, useCoords = TRUE, xlab = NULL, textSize = 1, ...)
}
\arguments{
\item{x}{A sites file object as read by readSites().}

\item{file}{(Alternative to x) Sites file (may be gzipped. may NOT be streamed)}

\item{start}{The start coordinate of the plot. NULL implies start from beginning of region}

\item{end}{The end coordinate of the plot. NULL implies go to end of region.}

\item{majColor}{The color of the major allele}

\item{minColor}{The color of the minor allele}

\item{stretch}{A horizontal expansion factor for sites (so that sparse sites are
visible in a large region)}

\item{indOrder}{A character vector giving order to list haplotypes (if NULL use
order used in sites object)}

\item{textColor}{a list saying which color to use for which haplotype label}

\item{useCoords}{if FALSE, just plot all variant sites with equal spacing}

\item{xlab}{label for x-axis. Default is "Coordinate" if useCoords=TRUE
or "Sites index" if useCoords=FALSE}

\item{textSize}{the magnification size of haplotype labelscat}

\item{...}{Other arguments passed to plot function}
}
\description{
Plot a sites file
}
