% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Plot statistics trace vs iteration}
\usage{
plotStats(x, stat, add = FALSE, xlim = NULL, ylim = NULL,
  xlab = "Iter", ylab = stat, ...)
}
\arguments{
\item{x}{A data.frame read in by readStatsFile. Alternatively, if
x is type character, can be the name of the stats file.}

\item{stat}{A character string naming the column of x to plot}

\item{add}{If FALSE, create a new plot. Otherwise add to current plot}

\item{xlim}{The x-axis range (only used when add==FALSE).
Taken from the range of data if not given.}

\item{ylim}{The y-axis range (only used when add==FALSE).
Taken from the range of data if not given}

\item{xlab}{The label for x-xais (only used when add=FALSE)}

\item{ylab}{The label for y-axis (only used when add=FALSE)}

\item{...}{Other options to pass to plot function}
}
\description{
Given a data frame read in by readStatsFile, plot one of the
columns vs iteration number.
}
