% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTraces.R
\name{plotTraces}
\alias{plotTraces}
\title{Plot ARGweaver statistics vs iteration}
\usage{
plotTraces(files = NULL, dir = NULL, outroot = NULL, stats = NULL,
  colors = if (length(files) == 1) {     "black" } else {    
  rainbow(length(files)) }, xlim = NULL, callPar = TRUE, ncol = 2,
  addLegend = (length(files) != 1), ...)
}
\arguments{
\item{files}{A character vector of file names giving the stat files to plot}

\item{dir}{(alternative to files; ignored if files is given). Plot all stat files found in this directory.}

\item{outroot}{(only used when dir is provided). Only match stat files named <outroot>.stat}

\item{stats}{A list of statistics to plot (i.e., "likelihood", "recombs", "prior", etc).
If NULL, then plot all statistics found in the stat files}

\item{colors}{A list of colors, one per file, to use in the plots}

\item{xlim}{The range of iterations to plot. If NULL, use entire range}

\item{callPar}{If TRUE, call the par() function to create a new graphics object with
a plot for each statistic on the same page}

\item{ncol}{(Only used if callPar=TRUE), arrange the plots in this many columns}

\item{addLegend}{If TRUE, add an additional plot that contains a legend mapping files to colors}

\item{...}{Arguments passed to plotStats()}
}
\value{
invisibly returns a data.frame detailing the files and colors used
}
\description{
Plots the statistics in stat files output by arg-sample vs iteration.
By default plots all statistics given in the stat file (as well
as any extra created by readStatFiles).
}
