% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTree}
\alias{plotTree}
\title{Plot a tree}
\usage{
plotTree(tree, prune = NULL, keepSeqs = NULL, col = "black",
  leafCol = col, leafLabels = NULL, timeScale = 1, drawSpr = FALSE,
  sites = NULL, chromStart = NULL, chromEnd = NULL,
  ylab = "Generations", logScale = FALSE, ylim = NULL, add = FALSE,
  mar = c(8, 4, 1, 1), ...)
}
\arguments{
\item{tree}{A newick string containing a tree}

\item{prune}{A list of leafs to prune from the tree before plotting (Null means prune none)}

\item{keepSeqs}{A list of leafs to keep in the tree (NULL means keep all)}

\item{col}{Either a character string (single value) giving color of tree. Or, a
list giving color to plot each leaf branch. Internal branches will
be plotted as "average" color among child leafs.}

\item{leafCol}{If given, this can be a list assigning a color to each leaf name
(individual names also work if haploid names are in the form indName_1 and indName_2)}

\item{leafLabels}{A list indexed by the leaf names in the tree, giving the label to display
for each leaf name. If NULL, the leaf names are displayed.}

\item{timeScale}{multiply all branches by this value}

\item{drawSpr}{If TRUE, draw the SPR event that turns each tree into the next one}

\item{sites}{If given, draw mutation events on tree}

\item{chromStart}{The start coordinate of the tree region (1-based; only used if SITES not null)}

\item{chromEnd}{The end coordinate of the tree region}

\item{ylab}{label for y axis}

\item{logScale}{If TRUE, plot in log scale}

\item{ylim}{Range for y axis (default; use range of tree)}

\item{add}{If TRUE, do not create a new plot}

\item{mar}{Margins for plot}

\item{...}{Passed to plot function}
}
\description{
Plot a tree
}
\note{
This creates a new plot for each tree. If plotting to the screen, probably
want to call par(ask=TRUE) first.
}
