% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{plotTrees}
\alias{plotTrees}
\title{Plot trees from newick string}
\usage{
plotTrees(trees, prune = NULL, keepSeqs = NULL, treeInfo = NULL,
  col = "black", leafCol = col, leafLabels = NULL, timeScale = 1,
  drawSpr = FALSE, ylab = "Generations", xlab = "", logScale = FALSE,
  ylim = NULL, add = FALSE, mar = c(8, 4, 1, 1), regionSide = 1,
  regionLine = 4, mod = NULL, popwidth = NULL, ...)
}
\arguments{
\item{trees}{A character vector, each element is a newick tree to plot}

\item{prune}{A list of leafs to prune from the tree before plotting (Null means prune none)}

\item{keepSeqs}{A list of leafs to keep in the tree (NULL means keep all)}

\item{treeInfo}{If given, should be a character vector of same length as trees}

\item{col}{Either a character string (single value) giving color of tree. Or, a
list giving color to plot each leaf branch. Internal branches will
be plotted as "average" color among child leafs.}

\item{leafCol}{If given, this can be a list assigning a color to each leaf name
(individual names also work if haploid names are in the form indName_1 and indName_2)}

\item{leafLabels}{A list indexed by the leaf names in the tree, giving the label to display
for each leaf name. If NULL, the leaf names are displayed.}

\item{timeScale}{multiply all branches by this value}

\item{drawSpr}{If TRUE, draw the SPR event that turns each tree into the next one}

\item{ylab}{label for y axis}

\item{xlab}{Label for x axis}

\item{logScale}{If TRUE, plot in log scale}

\item{ylim}{Range for y axis (default; use range of tree)}

\item{add}{If TRUE, do not create a new plot}

\item{mar}{Margins for plot}

\item{regionSide}{Print treeInfo to the margins on this side
(1=bottom, 2=left, 3=top, 4=right, anything else = don't print)}

\item{regionLine}{Print region of eaach tree on this line of the margin}

\item{mod}{(Advanced; for use with multi-population version of ARGweaver) A model
file read in with the function readPopModel, will draw population model underneath tree}

\item{popwidth}{If mod is not null, popwidth can be a numeric vector
of length equal to the number of populations, giving relative width
of each}

\item{...}{Passed to plot function}
}
\description{
Plot trees from newick string
}
\note{
This creates a new plot for each tree. If plotting to the screen, probably
want to call par(ask=TRUE) first.
}
