% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSites.R
\name{plotTreesAtSites}
\alias{plotTreesAtSites}
\title{call plotTreeFromBed at the sites in a site file; color individual nodes by allele}
\usage{
plotTreesAtSites(bedFile, sites, pos = NULL, start = NULL,
  end = NULL, doSingletons = FALSE, colors = list(A = "red", C =
  "green", G = "turquoise", T = "purple", N = "gray"), ...)
}
\arguments{
\item{bedFile}{bed.gz file from smc2bed with trees}

\item{sites}{Sites object, as returned by readSites(). Will plot one tree for each site.}

\item{pos}{If not NULL, a vector of integers giving which sites to plot.
If NULL, plot all sites}

\item{start}{start position}

\item{end}{end position}

\item{doSingletons}{whether to plot singletons}

\item{colors}{Color to use for each allele in leaf of tree}

\item{...}{Additional arguments passed to plotTreesFromBed}
}
\description{
call plotTreeFromBed at the sites in a site file; color individual nodes by allele
}
