% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{postorder.arg}
\alias{postorder.arg}
\title{Node postorder}
\usage{
postorder.arg(edge, leafOrder = NULL, mod = NULL)
}
\arguments{
\item{edge}{The edge matrix of a tree read by ape}

\item{leafOrder}{If given, this character vector determines how to sort child nodes}

\item{mod}{A population model for multi-population use}
}
\description{
Node postorder
}
