% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrees.R
\name{postorderRec}
\alias{postorderRec}
\title{Node postorder}
\usage{
postorderRec(edge, mod, idx, leafOrder = NULL)
}
\arguments{
\item{edge}{The edge matrix of a tree read by ape}

\item{mod}{A population model for multi-population use}

\item{idx}{The current idx}

\item{leafOrder}{If given, this character vector determines how to sort child nodes}
}
\description{
Node postorder
}
\note{
This is a recursive helper function for postorder.arg
}
\seealso{
\code{\link{postorder.arg}}
}
