% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmc.R
\name{psmcToPopsize}
\alias{psmcToPopsize}
\title{Make a popsize file suitable for arg-sample from PSMC output}
\usage{
psmcToPopsize(psmcFile, ntimes = 20, delta = 0.01, maxTime = 2e+05,
  mu = 1.4e-08, s = 100, sampleAge = 0, outfile = NULL)
}
\arguments{
\item{psmcFile}{File with PSMC output}

\item{ntimes}{Number of time points in arg-sample model}

\item{delta}{arg-sample delta parameter}

\item{maxTime}{maximum time point used in arg-sample}

\item{mu}{mutation rate in mutations per base pair per generation}

\item{s}{The bin size used in the fq2psmcfa command; the default in that program is 100}

\item{sampleAge}{The age of the sample, in generations (For ancient samples)}

\item{outfile}{If not NULL, path to write output popsize file. This file is suitable
for use with the "--popsize-file" option in arg-sample.}
}
\value{
a data frame with time and popsize (invisibly if outfile is provided)
}
\description{
Make a popsize file suitable for arg-sample from PSMC output
}
