% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readArgSummary.R
\name{readArgSummary}
\alias{readArgSummary}
\title{Read arg-sample summary file}
\usage{
readArgSummary(file, burnin = NULL)
}
\arguments{
\item{file}{A character string, either the name of the stats file or
the directory where the stat file can be found.
If a directory, there must be only one
stat file inside or else an error will occur}

\item{burnin}{If not NULL, a single numeric value. Any statistics from
iterations < burnin will be discarded. Will be ignored if there is not
a column named "MCMC_sample"}
}
\value{
A data frame containing the arg-summarize results
}
\description{
This function will read a file written by arg-summarize
}
