% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leafTrace.R
\name{readLeafTrace}
\alias{readLeafTrace}
\title{Read in a LeafTrace file}
\usage{
readLeafTrace(file)
}
\arguments{
\item{file}{The leaf trace file (produced by the script arg-layout)}
}
\value{
A data frame that can be set to plotLeafTrace
}
\description{
Read in a LeafTrace file
}
