% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popModel.R
\name{readPopFile}
\alias{readPopFile}
\title{Read in a population tree}
\usage{
readPopFile(popfile)
}
\arguments{
\item{popfile}{the file passed to arg-sample wtih --pop-tree-file}
}
\value{
A list with npop, divergence events (divs) and migration events (migs)
}
\description{
Read in a population tree
}
