% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popModel.R
\name{readPopModel}
\alias{readPopModel}
\title{Read in a population model file}
\usage{
readPopModel(logfile, popfile)
}
\arguments{
\item{logfile}{The log file output by arg-sample}

\item{popfile}{The population file given to arg-sample with the argument
--pop-tree-file}
}
\value{
A population object suitable to pass to plotTree function
}
\description{
Read in a population model file
}
