% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psmc.R
\name{readPsmc}
\alias{readPsmc}
\title{Read a PSMC output file}
\usage{
readPsmc(psmcFile, mu = 1.4e-08, s = 100, sampleAge = 0)
}
\arguments{
\item{psmcFile}{file name containing PSMC output}

\item{mu}{Mutation rate per base pair per generation}

\item{s}{The site compression rate used in PSMC}

\item{sampleAge}{(For ancient samples) the sample age, in generations}
}
\value{
A data frame containing fields minTime, maxTime, popsize. MinTime
and MaxTime are in generations and will be adjusted by sampleAge
}
\description{
Read a PSMC output file
}
