% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSites.R
\name{readSites}
\alias{readSites}
\title{Read an ARGweaver sites file}
\usage{
readSites(file)
}
\arguments{
\item{file}{Sites file (may be gzipped. may NOT be streamed)}
}
\value{
a list with two elements: one describing the region, and the
other a data frame with a column for each haploid genome and a row
for each variant site
}
\description{
Read an ARGweaver sites file
}
