% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popModel.R
\name{scalePopModel}
\alias{scalePopModel}
\title{scalePopModel: scale times in a population model}
\usage{
scalePopModel(mod, scale)
}
\arguments{
\item{mod}{A popmodel (usually read by readPopModel)}

\item{scale}{Scaling factor}
}
\value{
a new popmodel with scaled times
}
\description{
scalePopModel: scale times in a population model
}
