% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSites.R
\name{subSites}
\alias{subSites}
\title{Subset a sites object}
\usage{
subSites(x, keep = NULL, prune = NULL, region = NULL)
}
\arguments{
\item{x}{a sites object (read in by readSites)}

\item{keep}{A list of haplotypes to keep}

\item{prune}{A list of haplotypes to prune}

\item{region}{A numeric vector of length 2 giving minimum and maximum coordinates
(1-based)}
}
\value{
A sites object subsetted by haplotypes and/or region
}
\description{
Subset a sites object
}
