% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeSites.R
\name{writeSites}
\alias{writeSites}
\title{Write an ARGweaver sites file}
\usage{
writeSites(sites, filename)
}
\arguments{
\item{sites}{Sites object}

\item{file}{Sites file (may be gzipped. may NOT be streamed)}
}
\value{
a list with two elements: one describing the region, and the
other a data frame with a column for each haploid genome and a row
for each variant site. A third element "probs" will be added if readProbs=TRUE
}
\description{
Write an ARGweaver sites file
}
