% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBw.R
\name{plotBw}
\alias{plotBw}
\title{plot bigWig (a.k.a. bedGraph-style) data}
\usage{
plotBw(x, y, y0 = NULL, y1 = NULL, col = "black", alpha = 0.2,
  add = FALSE, xlim = NULL, chrom = NULL, ylab = "Generations",
  xlab = "Coordinate", ylim = NULL, ...)
}
\arguments{
\item{x}{a data frame with at least three columns with names
chrom, chromStart, chromEnd}

\item{y}{a numeric vector whose length=nrow(x) giving one score per
region represented in x}

\item{y0}{A lower limit to y (should be numeric vector, same length as x and y)}

\item{y1}{An upper limit to y (same format as y0; must be provided if y0 is provided)}

\item{col}{The color of the line to draw}

\item{alpha}{Transparency factor (0 = transparent, 1 = solid)}

\item{add}{If TRUE, add to current plot}

\item{xlim}{The x-axis limits}

\item{chrom}{If given plot only elements of x for each x$chrom==chrom}

\item{ylab}{label for y-axis}

\item{xlab}{label for x-axis}

\item{ylim}{limits for y-axis. If not given will use range(y)}

\item{...}{Other arguments to be passed to plot functions

If y0 and y1 are given, then will shade in regions between y0 and y1 in the same
color as the line drawn, but made transparent by factor "alpha'
#TODO: have it take an optional low/high value, as well as optional alpha}
}
\description{
plot bigWig (a.k.a. bedGraph-style) data
}
